Refactor MessagePartView. Removed ViewPart trait, not really needed, we can use MessageViewPart everywhere. Moved view part definitions to separate file.
