+++ app/src/main/res/layout/message_ephemeral_dots_view.xml
@@ -1,0 +1,34 @@
+<?xml version="1.0" encoding="utf-8"?>
+<!--
+
+    Wire
+    Copyright (C) 2016 Wire Swiss GmbH
+
+    This program is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    This program is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with this program.  If not, see <http://www.gnu.org/licenses/>.
+
+-->
+<com.waz.zclient.ui.views.EphemeralDotAnimationView
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
+    android:id="@+id/edav__ephemeral_view"
+    android:layout_width="@dimen/ephemeral__animating_dots__width"
+    android:layout_height="@dimen/ephemeral__animating_dots__height"
+    android:layout_gravity="end|top"
+    android:layout_marginEnd="@dimen/ephemeral__animating_dots__right_margin"
+    android:paddingTop="@dimen/ephemeral__animating_dots__padding_top"
+    android:paddingBottom="@dimen/ephemeral__animating_dots__padding_bottom"
+    app:primaryColor="@color/ephemera"
+    app:secondaryColor="@color/ephemera_dimmed"
+    app:dotPadding="@dimen/ephemeral__animating_dots__space_between"
+/>
